Setup LazyVim initialization
