--- roles/neovim/files/after/ftplugin/css.vim
@@ -1,1 +1,0 @@
-setlocal tabstop=2 shiftwidth=2 expandtab
