--- roles/neovim/files/after/ftplugin/javascript.vim
@@ -1,1 +1,0 @@
-setlocal tabstop=2 shiftwidth=2 expandtab
