--- roles/neovim/files/after/ftplugin/python.vim
@@ -1,1 +1,0 @@
-setlocal tabstop=4 shiftwidth=4 expandtab
