--- roles/neovim/files/after/ftplugin/sh.vim
@@ -1,1 +1,0 @@
-setlocal tabstop=2 shiftwidth=2 expandtab
