--- roles/neovim/files/after/plugin/colors.lua
@@ -1,15 +1,0 @@
-require("catppuccin").setup({
-    flavour = "mocha", -- latte, frappe, macchiato, mocha
-    transparent_background = true,
-})
-
-function ColorMyPencils(color)
-	color = color or "catppuccin"
-    vim.cmd.colorscheme(color)
-
-	vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
-	vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
-
-end
-
-ColorMyPencils()
