--- roles/neovim/files/after/plugin/trouble.lua
@@ -1,3 +1,0 @@
-vim.keymap.set("n", "<leader>xq", "<cmd>TroubleToggle quickfix<cr>",
-  {silent = true, noremap = true}
-)
