+++ roles/neovim/files/lua/config/init.lua
@@ -1,5 +1,9 @@
 require("config.remap")
 require("config.set")
+
+require("config.lazy_init")
+
+--[[
 require("config.packer")
 
 local augroup = vim.api.nvim_create_augroup
@@ -23,12 +27,14 @@ autocmd('TextYankPost', {
     end,
 })
 
-autocmd({"BufWritePre"}, {
-    group = ConfGroup,
-    pattern = "*",
-    command = [[%s/\s\+$//e]],
-})
+]]--
 
-vim.g.netrw_browse_split = 0
-vim.g.netrw_banner = 0
-vim.g.netrw_winsize = 25
+--autocmd({"BufWritePre"}, {
+--    group = ConfGroup,
+--    pattern = "*",
+--    command = [[%s/\s\+$//e]],
+--})
+--
+--vim.g.netrw_browse_split = 0
+--vim.g.netrw_banner = 0
+--vim.g.netrw_winsize = 25
