+++ roles/neovim/files/lua/config/lazy/colors.lua
@@ -1,0 +1,19 @@
+function ColorMyPencils(color)
+	color = color or "catppuccin"
+	vim.cmd.colorscheme(color)
+
+	vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
+	vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
+
+end
+
+return {
+    {
+        "catppuccin/nvim",
+        lazy = false,
+        priority = 1000,
+        config = function() 
+            ColorMyPencils()
+        end,
+    },
+}
